histogram integer (tests broken) #683
