+++ server/opendp_apps/analysis/testing/__init__.py
@@ -3,6 +3,8 @@ docker-compose run server python manage.py test opendp_apps.analysis.testing.tes
 
 docker-compose run server python manage.py test opendp_apps.analysis.testing.test_dp_mean_spec.StatSpecTest.test_40_test_impute
 
+docker-compose run server python manage.py test opendp_apps.analysis.testing.test_dp_histogram_integer_spec.HistogramIntegerStatSpecTest.test_120_run_dphist_calculation_integer
+
 python manage.py test opendp_apps.analysis.testing.test_dp_count_spec.DPCountStatSpecTest.test_10_valid_spec
 
 python manage.py test opendp_apps.analysis.testing.test_run_release.TestRunRelease.test_10_compute_stats
